Implement serialization for SpendableOutput
Changes the output index to a u8. While it may expand to a u16 at some 
point, this can remain canonical using little endian serialization while 
dropping the latter byte if it's 0 (or simply only using u16 when it's 
actually possible).

+++ coins/monero/src/wallet/decoys.rs
@@ -110,7 +110,7 @@ impl Decoys {
     let mut outputs = Vec::with_capacity(inputs.len());
     for input in inputs {
       outputs.push((
-        rpc.get_o_indexes(input.tx).await?[input.o],
+        rpc.get_o_indexes(input.tx).await?[usize::from(input.o)],
         [input.key, input.commitment.calculate()]
       ));
     }
